Change connected components data type to uint16 (#1763)
* Allocate connected components as uint16

* Change connected component invalid value in CUDA geocode

* support unsigned short/GDT_UInt16

* more explicit instantiations
reformat explicit instantiation for readability

* Remove hard-coded fill value for connect components

* Fill values for unsigned and signed integers

---------

+++ python/packages/nisar/products/insar/GUNW_writer.py
@@ -243,7 +243,7 @@ class GUNWWriter(RUNWWriter, RIFGWriter, L2InSARWriter):
                     ("coherenceMagnitude", np.float32,
                      f"Coherence magnitude between {pol} layers",
                      Units.unitless),
-                    ("connectedComponents", np.uint32,
+                    ("connectedComponents", np.uint16,
                      f"Connected components for {pol} layers",
                      Units.dn),
                     ("ionospherePhaseScreen", np.float32,
